requirements.txt
Flask==3.1.1
requests==2.31.0
geopy==2.4.1
polyline==2.0.0   
python-dotenv==1.0.1
